requirements.txt
openai
python-dotenv
cycode
e2b-code-interpreter
requests
flask
pytest
